pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "radbot"
version = "0.1.0"
authors = [
    {name = "Perry Manuk", email = "perry.manuk@example.com"},
]
description = "A modular AI agent framework using Google ADK, Qdrant, MCP, and A2A"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "google-adk>=1.22.0",
    "google-genai>=1.57.0",
    "qdrant-client>=1.13.0,<1.17",
    "modelcontextprotocol>=0.1.0",
    "zoneinfo; python_version < '3.9'",
    "tzdata",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "nest-asyncio>=1.6.0",  # Required for MCP fileserver to work in running event loops
    "requests>=2.31.0",
    "websockets>=12.0.0",   # Required for Home Assistant WebSocket client
    "httpx>=0.27.0",        # HTTP client for async requests
    "psycopg2-binary>=2.9.9",  # Required for PostgreSQL database connections
    "redis>=5.0.1",         # Optional for cross-session caching with Redis
    "google-api-python-client>=2.102.0", # Required for Google Calendar API
    "google-auth-httplib2>=0.1.0",      # Required for Google Calendar API authentication
    "google-auth-oauthlib>=1.1.0",      # Required for Google Calendar OAuth flow
    "python-jsonrpc>=0.10.0",          # Required for Context7 MCP stdio client
    "apscheduler>=3.10.0",             # Required for cron/scheduling engine
    "google-cloud-texttospeech>=2.14.0",  # Required for TTS synthesis
    "atlassian-python-api>=3.41.0",        # Required for Jira Cloud integration
    "cryptography>=42.0.0",                # Required for credential store encryption
    # Removed gcsa due to dependency conflict with ADK (tzlocal version conflict)
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.12.0",
    "flake8>=7.0.0",
    "mypy>=1.5.0",
    "pytest>=8.3.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
]

web = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "jinja2>=3.1.4",
    "starlette>=0.36.0",
    "python-multipart>=0.0.9",
]

[project.urls]
"Homepage" = "https://github.com/perrymanuk/radbot"
"Bug Tracker" = "https://github.com/perrymanuk/radbot/issues"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tools", "examples", ".git", ".venv", "node_modules", "build", "dist"]
collect_ignore_glob = ["tools/*", "examples/*"]